subscription_cost/requirements.txt
asgiref==3.8.1
Django==4.2.19
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.15.2
django-stubs[compatible-mypy]==4.2.*
django-extensions==3.2.3
ipython==8.18.1
python-dateutil==2.9.0
django-cors-headers == 4.7.0
celery==5.3.6

